repair int to size_t warn

+++ leetcode/0-1000/0/3_无重复字符的最长子串.cc
@@ -12,7 +12,7 @@ int lengthOfLongestSubstring(string s) {
         chars.insert(s[right++]);
         max = std::max(max, right - left);
     }
-    return max;
+    return static_cast<int>(max);
 }
 
 TEST(solution, lengthOfLongestSubstring) {
